requirements.txt
html2text
markdownify
pdfminer
python-docx
openpyxl
requests
beautifulsoup4
tkinterdnd2
